#354 Initial Refactoring
